# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API MathEasy",
        "description": "API per gestionar operacions i recursos de MathEasy amb Laravel i autenticació Bearer",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://192.168.1.39:8000/api",
            "description": "Servidor local de desenvolupament"
        }
    ],
    "paths": {
        "/alumnes": {
            "get": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Llistar tots els alumnes",
                "description": "Retorna una llista amb tots els alumnes registrats",
                "operationId": "136a14845978f96f58e89b302a2add0e",
                "responses": {
                    "200": {
                        "description": "Llista d'alumnes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Alumne"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Crear un nou alumne",
                "description": "Crea un alumne nou amb les dades proporcionades",
                "operationId": "f402569706d709e193d14dd37fabd227",
                "requestBody": {
                    "description": "Dades de l'alumne",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlumneInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alumne creat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alumne"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dades no vàlides"
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alumnes/{id}": {
            "get": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Mostrar un alumne",
                "description": "Retorna les dades d'un alumne específic pel seu ID",
                "operationId": "9080ee90af4f7973038fd9b97d773847",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'alumne",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alumne trobat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alumne"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alumne no trobat"
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Actualitzar dades d'un alumne",
                "description": "Actualitza les dades personals d'un alumne autenticat.",
                "operationId": "d090b6e280b0588cc1e834e7f425edb9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'alumne",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dades a actualitzar de l'alumne",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlumneUpdateInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alumne actualitzat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alumne"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticat"
                    },
                    "404": {
                        "description": "Alumne no trobat"
                    },
                    "422": {
                        "description": "Dades no vàlides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/loggin": {
            "post": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Iniciar sessió d'un alumne",
                "description": "Comprova el nom d'usuari i la contrasenya. Retorna les dades de l'alumne si és correcte.",
                "operationId": "eed7e793b3ea10acda6b6bd66e2a2c56",
                "requestBody": {
                    "description": "Dades d'inici de sessió",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuari autenticat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuari i/o contrasenya incorrectes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tokenLoggin": {
            "post": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Validar token",
                "operationId": "df7f3482c82eba4de12be93d284b8baa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TokenLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token vàlid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alumne"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token invàlid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/perfilImage": {
            "post": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Pujar imatge de perfil",
                "operationId": "4965a80c46031b213507fe005960509c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileImageUpload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Imatge pujada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/perfilImageEdit": {
            "post": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Editar imatge de perfil",
                "operationId": "b331dbb0973bde12ef495a85d840b8bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileImageEdit"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Imatge actualitzada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alumnes/experiencia/{id}": {
            "get": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Obtenir experiència",
                "description": "Retorna dades d'experiència d'un alumne",
                "operationId": "b012d53cd33bdbf8515f7f26b41c9d23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Experiència trobada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Experiencia"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No trobada"
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Actualitzar experiència",
                "description": "Actualitza nivell, XP i medalles",
                "operationId": "4945977539b9bec25501429e3271be05",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'experiencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExperienciaUpdateInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualitzat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Experiencia"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticat"
                    },
                    "404": {
                        "description": "Alumne no trobat"
                    },
                    "422": {
                        "description": "Dades no vàlides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/informes": {
            "get": {
                "tags": [
                    "Informes"
                ],
                "summary": "Llistar tots els informes",
                "description": "Retorna una llista amb tots els informes creats",
                "operationId": "427188f4b0e442f4d4a2bf1f2ff7c66f",
                "responses": {
                    "200": {
                        "description": "Llista d'informes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Informe"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Informes"
                ],
                "summary": "Crear un nou informe",
                "description": "Crea un informe amb les dades proporcionades",
                "operationId": "5913480bd8bc3db0d68df030c3ea9bb2",
                "requestBody": {
                    "description": "Dades de l'informe",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InformeInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Informe creat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Informe"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dades no vàlides"
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/informesAlumne/{alumne}": {
            "get": {
                "tags": [
                    "Informes"
                ],
                "summary": "Llistar tots els informes d'un alumne concret",
                "description": "Retorna una llista amb tots els informes creats d'un alumne concret",
                "operationId": "32de57bfe360d0c94eca3faf569c6683",
                "parameters": [
                    {
                        "name": "alumne",
                        "in": "path",
                        "description": "ID de l'alumne",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Llista d'informes d'un alumne concret",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Informe"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alumne no trobat"
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/professors": {
            "post": {
                "tags": [
                    "Professors"
                ],
                "summary": "Crear un nou professor",
                "description": "Crea un professor nou amb les dades proporcionades",
                "operationId": "ae9c2e9a3dc3e4f81e747d4429ccb1cb",
                "requestBody": {
                    "description": "Dades de l'alumne",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfessorInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Professor creat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Professor"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dades no vàlides"
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "AlumneInput": {
                "title": "Entrada Alumne",
                "description": "Dades necessàries per crear un alumne",
                "properties": {
                    "Nom": {
                        "description": "Nom de l'usuari",
                        "type": "string"
                    },
                    "Cognoms": {
                        "description": "Cognoms de l'usuari",
                        "type": "string"
                    },
                    "Password": {
                        "description": "Password de l'usuari",
                        "type": "string",
                        "format": "password"
                    },
                    "Nom_Usuari": {
                        "description": "Nom d'usuari dins del curs",
                        "type": "string"
                    },
                    "Curs": {
                        "description": "Curs de l'alumne",
                        "type": "string"
                    },
                    "Experiencia": {
                        "description": "Experiència de l'alumne",
                        "type": "integer",
                        "format": "int32"
                    },
                    "ProfilePicturePath": {
                        "description": "Imatge de perfil",
                        "type": "string",
                        "example": "images/avatar.jpg"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduïu el token Bearer JWT per autenticar-vos",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}